WIP: Implement search filter api and db layer

+++ models/search.go
@@ -55,7 +55,7 @@ type SearchFilter struct {
 type SavedSearchFilter struct {
 	SearchFilter
 
-	ID     int    `json:"id" db:"id"`
-	UserID int    `json:"userId" db:"user_id"`
+	ID     int64  `json:"id" db:"id"`
+	UserID int64  `json:"userId" db:"user_id"`
 	Name   string `json:"name" db:"name"`
 }
